Updated IO subsystem the match changes done in nboxkrnl

+++ src/io.hpp
@@ -13,4 +13,4 @@ bool io_init(const char *nxbx_path, const char *xbe_path);
 void io_stop();
 void submit_io_packet(uint32_t addr);
 void flush_pending_packets();
-uint32_t query_io_packet(uint32_t id, bool query_status);
+uint32_t query_io_packet(uint64_t id, bool query_status);
